+++ Servidor/src/vo/Gasto.java
@@ -18,7 +18,7 @@ public class Gasto implements Serializable {
     private int id;
     private String nombre_gasto;
     private String tipo_gasto;
-    private int precio_gasto = 0;
+    private double precio_gasto = 0.0;
 
     /**
      * Constructor por defecto de la clase Gasto
@@ -98,7 +98,7 @@ public class Gasto implements Serializable {
      *
      * @return precio del gasto
      */
-    public int getPrecio_gasto() {
+    public double getPrecio_gasto() {
         return precio_gasto;
     }
 
@@ -107,7 +107,7 @@ public class Gasto implements Serializable {
      *
      * @param precio_gasto precio del gasto
      */
-    public void setPrecio_gasto(int precio_gasto) {
+    public void setPrecio_gasto(double precio_gasto) {
         this.precio_gasto = precio_gasto;
     }
